+++ changelog.md
@@ -8,6 +8,10 @@ Date format is `MM/DD/YYYY`.
 ### Added
 
 - getCurrentUserRoles() method to get the current user's roles
+- changelog
+- getSelf() method to get the current user's information
+- getUsers() method to get a list of users via IDs
+- getUsersByUserNames() method to get a list of users via usernames
 
 ## [0.0.1] - 1/14/2024